requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
albumentations @ git+https://github.com/albu/albumentations@81523ead5bf0dab7b8396c3e63230d317a768683
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
fsspec==2021.5.0
future==0.18.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
imageio==2.9.0
install==1.3.4
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
multidict==5.1.0
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.1
opencv-python-headless==4.5.2.52
packaging==20.9
Pillow==8.2.0
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.3.3
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.4.8
timm==0.4.9
torch==1.8.1
torchmetrics==0.3.2
torchvision==0.9.1
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
